Editor separated to widget Fix error with select connection after another was initiated
