+++ default.css
@@ -9,7 +9,7 @@ TabbedContent {
 }
 
 DataTable {
-height: 100%;
+    height: 100%;
 }
 
 TabPane {
